\name{automobile}
\alias{automobile}
\title{Data concerning automobile prices.}
\description{
  The \code{automobile} data frame has 160 rows and 12 columns and
  concerns white wines from a region in Portugal. The response 
  variable, \code{quality}, is a wine quality metric, taken to be the median
 preference score of three blind tasters on a scale of 1-10. The 11 covariates
 are physicochemical metrics of wine quality such as citric acid content, 
 sulphates, etc.
}
\usage{
WhiteWine
}
\source{
K Bache and M Lichman. UCI machine learning repository, 2013. 
http://archive.ics.uci.edu/ml
}
\keyword{datasets}
