% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bart_package_inits.R
\name{bart_machine_num_cores}
\alias{bart_machine_num_cores}
\title{Get Number of Cores Used by BART}
\usage{
bart_machine_num_cores()
}
\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
Number of cores currently being used by parallelized BART functions.
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
\description{
Returns number of cores used by BART
}
\details{
Returns the number of cores currently being used by parallelized BART functions
}
\examples{
\dontrun{
bart_machine_num_cores()
}
}
\seealso{
\code{\link{set_bart_machine_num_cores}}
}
\author{
Adam Kapelner and Justin Bleich
}
