% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bartMachine-package.R
\docType{package}
\name{bartMachine-package}
\alias{bartMachine-package}
\title{bartMachine: Bayesian Additive Regression Trees}
\description{
An advanced implementation of Bayesian Additive Regression Trees with expanded features for data analysis and visualization.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/kapelner/bartMachine/issues}
}

}
\author{
\strong{Maintainer}: Adam Kapelner \email{kapelner@qc.cuny.edu} (\href{https://orcid.org/0000-0001-5985-6792}{ORCID})

Authors:
\itemize{
  \item Justin Bleich
}

}
\keyword{internal}
