% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bart_package_builders.R
\name{destroy_bart_machine}
\alias{destroy_bart_machine}
\title{Destroy BART Model (deprecated --- do not use!)}
\usage{
destroy_bart_machine(bart_machine)
}
\arguments{
\item{bart_machine}{deprecated --- do not use!}
}
\value{
None.
}
\description{
A deprecated function that previously was responsible for cleaning up the RAM
associated with a BART model. This is now handled natively by R's garbage collection.
}
\details{
Removing a ``bart_machine'' object from \code{R} previously did not free heap space from Java.
Since BART objects can consume a large amount of RAM, it is important to remove
these objects by calling this function if they are no longer needed or many BART
objects are being created. This operation is now taken care of by R's garbage collection.
This function is deprecated and should not be used. However, running it is harmless.
}
\examples{
##None
}
\author{
Adam Kapelner and Justin Bleich
}
