\name{plot_convergence_diagnostics}
\alias{plot_convergence_diagnostics}
\title{no title}
\description{
  no description
}
\usage{
  plot_convergence_diagnostics()
}
\details{
  no details
}
\value{
	no value
}
\author{
	Adam Kapelner and Justin Bleich
}
\examples{
  plot_convergence_diagnostics()
}