% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bart_package_inits.R
\name{set_bart_machine_num_cores}
\alias{set_bart_machine_num_cores}
\title{Set the Number of Cores for BART}
\usage{
set_bart_machine_num_cores(num_cores)
}
\arguments{
\item{num_cores}{Number of cores to use. If the number of cores is more than 1, setting the seed during model construction
cannot be deterministic.}
}
\value{
None.
}
\description{
Sets the number of cores to be used for all parallelized BART functions.
}
\examples{
\dontrun{
#set all parallelized functions to use 4 cores
set_bart_machine_num_cores(4)
}
}
\seealso{
\code{\link{bart_machine_num_cores}}
}
\author{
Adam Kapelner and Justin Bleich
}
