\name{bart_machine_num_cores}
\alias{bart_machine_num_cores}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Get Number of Cores Used by BART
}
\description{
Returns number of cores used by BART
}
\usage{
bart_machine_num_cores()
}
%- maybe also 'usage' for other objects documented here.
\details{
Returns the number of cores currently being used by parallelized BART functions
}
\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
Number of cores currently being used by parallelized BART functions.
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
Adam Kapelner and Justin Bleich
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{set_bart_machine_num_cores}}
}
\examples{
\dontrun{
bart_machine_num_cores()
}
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
