\name{benchmark_datasets}
\alias{benchmark_datasets}
\title{Benchmark Datasets}
\description{
  Nine eclectic datasets which were used for benchmarking bartMachine's out of sample performance in 
  the vignette for this package.
}
\usage{
automobile
}
\source{
K Bache and M Lichman. UCI machine learning repository, 2013. 
http://archive.ics.uci.edu/ml/datasets/Automobile
}
\references{
A Kapelner and J Bleich. bartMachine: a Powerful Tool for Machine Learning. ArXiv e-prints, 2013.
}
\keyword{datasets}
