% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{benchmark_datasets}
\alias{benchmark_datasets}
\alias{ankara}
\alias{baseball}
\alias{boston}
\alias{compactiv}
\alias{ozone}
\alias{pole}
\alias{triazine}
\alias{wine.red}
\alias{wine.white}
\title{benchmark_datasets}
\source{
See vignette for details.
}
\usage{
data(benchmark_datasets)
}
\description{
Nine diverse datasets which were used for benchmarking bartMachine's out of sample performance in
  the vignette for this package.
}
\keyword{datasets}
