\name{check_bart_error_assumptions}
\alias{check_bart_error_assumptions}
\title{no title}
\description{
  no description
}
\usage{
  check_bart_error_assumptions()
}
\details{
  no details
}
\value{
	no value
}
\author{
	Adam Kapelner and Justin Bleich
}
\examples{
  check_bart_error_assumptions()
}