\name{var_selection_by_permute_response_three_methods}
\alias{var_selection_by_permute_response_three_methods}
\title{no title}
\description{
  no description
}
\usage{
  var_selection_by_permute_response_three_methods()
}
\details{
  no details
}
\value{
	no value
}
\author{
	Adam Kapelner and Justin Bleich
}
\examples{
  var_selection_by_permute_response_three_methods()
}